Enable MTU size to be configured
Add a mtu flag on the main garden-linux executable and pass its value into the
Linux backend and thence to each container's Start method. Set the MTU size of
both ends of the container's virtual ethernet pair to the specified value.

Syntax:

-mtu=<uint32>

Default value: 1500

In addition to unit tests, add an integration test which checks the MTU size
is configured correctly on both ends of a container's virtual ethernet pair.

[#80221576]